refactor(codegen): extract return-path deep-copy into locus/return_path.rs (Round 7c)
5 fns lifted as inherent impls on Cx into locus/return_path.rs
(~860 lines):
- emit_return_value_deep_copy (the m49 / m90 method-return deep
  copy — at every locus-method return site, copy the success
  value out of the method's per-call scratch arena into the
  caller's arena before tearing scratch down)
- emit_cross_arena_store_deep_copy_ptr (the field-store
  pointer-typed variant)
- emit_enum_payload_deep_copy (variant-aware copy that walks the
  active variant's fields per the i32 tag)
- emit_self_field_inplace_assign (the `self.X = ...` store path
  with arena-boundary deep-copy when crossing arenas)
- maybe_self_field_heap_copy (the heap-copy branch used when
  self.X holds an arena-allocated value)

This is the missing 4.5 sub-round from the original Round 4 plan —
where issue #9 (m90 return-slot ABI) will localize. The whole
arena-boundary deep-copy story now lives in one file.

Surface bumps to pub(crate):
- Cx methods: emit_enum_payload_deep_copy, emit_self_field_inplace_assign,
  maybe_self_field_heap_copy, compound_storage_size,
  iface_fat_struct_ty, llvm_tuple_storage_type,
  load_enum_payload_fields, load_enum_tag, lower_enum_variant_alloc

codegen.rs: 20889 -> 20023 (-866 lines).
