+++ crates/hale-codegen/src/locus/mod.rs
@@ -6,3 +6,4 @@ pub(crate) mod decl;
 pub(crate) mod dissolve;
 pub(crate) mod instantiation;
 pub(crate) mod method;
+pub(crate) mod return_path;
